requirements.txt
streamlit>=1.40.0
groq>=1.0.0
chromadb>=1.5.0
pymupdf>=1.24.0
youtube-transcript-api>=0.6.0
numpy>=1.26.0
pandas>=2.0.0
altair>=5.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
